Refactor training plan repository structure by consolidating functions into a dedicated 'training-plans' directory. Implement new functions for creating, fetching, and updating training plans, enhancing organization and maintainability of the codebase.
