--- lib/repository/trainingPlan.ts
@@ -1,243 +1,0 @@
-import { createClient } from '@supabase/supabase-js';
-import type { TrainingPlan, CreateTrainingPlanInput } from "@/lib/types";
-import type { JSON } from "@/lib/types";
-// Import the shared Supabase service role client
-import { supabaseServiceRole } from "@/lib/supabase/serviceRoleClient";
-
-// Function to create a new training plan in the database
-export const createTrainingPlan = async (
-    userId: string | null,
-    data: CreateTrainingPlanInput & { overview: string; planJson: JSON },
-): Promise<TrainingPlan> => {
-    console.log("Saving training plan to database:", data);
-
-    // Map the input data to the database schema (assuming snake_case)
-    const { data: newTrainingPlan, error } = await supabaseServiceRole
-        .from('training_plans') // Assuming your table name is 'training_plans'
-        .insert([
-            {
-                client_id: data.clientId,
-                title: data.title,
-                overview: data.overview,
-                plan_json: data.planJson, // Store JSON data directly
-                assistant_ids: data.assistantIds, // Assuming these are stored as an array type or JSONB
-                goal_ids: data.goalIds, // Assuming these are stored as an array type or JSONB
-                // Supabase often auto-manages created_at, updated_at, id
-                // generated_by: data.generatedBy, // Optional: if you pass this from the resolver
-                // source_prompt: data.sourcePrompt, // Optional: if you pass this from the resolver
-            },
-        ])
-        .select()
-        .single(); // Select the newly inserted row
-
-    if (error) {
-        console.error("Error saving training plan:", error);
-        throw new Error(`Failed to create training plan: ${error.message}`);
-    }
-
-    // Map the database response back to the GraphQL TrainingPlan type (assuming camelCase in GraphQL)
-    // Note: You might need a more robust mapping function if column names differ significantly
-    const createdTrainingPlan: TrainingPlan = {
-        id: newTrainingPlan.id,
-        clientId: newTrainingPlan.client_id,
-        title: newTrainingPlan.title,
-        overview: newTrainingPlan.overview,
-        planJson: newTrainingPlan.plan_json,
-        assistantIds: newTrainingPlan.assistant_ids,
-        goalIds: newTrainingPlan.goal_ids,
-        createdAt: new Date(newTrainingPlan.created_at), // Assuming Supabase returns timestamp strings
-        updatedAt: new Date(newTrainingPlan.updated_at), // Assuming Supabase returns timestamp strings
-        deletedAt: newTrainingPlan.deleted_at ? new Date(newTrainingPlan.deleted_at) : null,
-        generatedBy: newTrainingPlan.generated_by || null, // Handle potential nulls
-        sourcePrompt: newTrainingPlan.source_prompt || null, // Handle potential nulls
-    };
-
-    console.log("Successfully saved training plan:", createdTrainingPlan);
-
-    return createdTrainingPlan;
-};
-
-// Function to update a training plan in the database
-export const updateTrainingPlan = async (
-    userId: string | null,
-    trainingPlanId: string,
-    data: { overview?: string; planJson?: JSON; generatedBy?: string | null; sourcePrompt?: string | null },
-): Promise<TrainingPlan | null> => {
-    console.log("Updating training plan", trainingPlanId, "with data:", data);
-
-    if (!userId) {
-        console.log('updateTrainingPlan: No user ID provided.');
-        return null;
-    }
-
-    // Ensure the training plan belongs to the user before updating
-    const existingPlan = await supabaseServiceRole
-        .from('training_plans')
-        .select('id')
-        .eq('id', trainingPlanId)
-        .eq('user_id', userId)
-        .single();
-
-    if (!existingPlan) {
-        console.log(`Training plan with ID ${trainingPlanId} not found for user ${userId}, cannot update.`);
-        return null;
-    }
-
-    const { data: updatedPlan, error } = await supabaseServiceRole
-        .from('training_plans')
-        .update({
-            overview: data.overview,
-            plan_json: data.planJson, // Store JSON data directly
-            generated_by: data.generatedBy, // Optional: if you pass this from the resolver
-            source_prompt: data.sourcePrompt, // Optional: if you pass this from the resolver
-            // updated_at will be automatically managed by Supabase if configured
-        })
-        .eq('id', trainingPlanId)
-        .select()
-        .single();
-
-    if (error) {
-        console.error(`Error updating training plan ${trainingPlanId}:`, error);
-        throw new Error(`Failed to update training plan: ${error.message}`);
-    }
-
-    if (!updatedPlan) {
-        console.error(`Update operation for training plan ${trainingPlanId} returned no data.`);
-        return null; // Should not happen if there's no error, but good practice
-    }
-
-    // Map the database response back to the GraphQL TrainingPlan type
-    const trainingPlan: TrainingPlan = {
-        id: updatedPlan.id,
-        clientId: updatedPlan.client_id,
-        title: updatedPlan.title,
-        overview: updatedPlan.overview,
-        planJson: updatedPlan.plan_json,
-        assistantIds: updatedPlan.assistant_ids,
-        goalIds: updatedPlan.goal_ids,
-        createdAt: new Date(updatedPlan.created_at), // Assuming Supabase returns timestamp strings
-        updatedAt: new Date(updatedPlan.updated_at), // Assuming Supabase returns timestamp strings
-        deletedAt: updatedPlan.deleted_at ? new Date(updatedPlan.deleted_at) : null,
-        generatedBy: updatedPlan.generated_by || null, // Handle potential nulls
-        sourcePrompt: updatedPlan.source_prompt || null, // Handle potential nulls
-    };
-
-    console.log("Successfully updated training plan:", trainingPlan);
-
-    return trainingPlan;
-};
-
-// Function to get a training plan by its ID
-export const getTrainingPlanById = async (id: string): Promise<TrainingPlan | null> => {
-    console.log("Fetching training plan by ID:", id);
-
-    const { data: trainingPlan, error } = await supabaseServiceRole
-        .from('training_plans')
-        .select('*')
-        .eq('id', id)
-        .single();
-
-    if (error) {
-        console.error(`Error fetching training plan with ID ${id}:`, error);
-        // Depending on error type, might throw or return null. Returning null for not found is common.
-        return null;
-    }
-
-    if (!trainingPlan) {
-        console.log(`Training plan with ID ${id} not found.`);
-        return null;
-    }
-
-    // Map the database response back to the GraphQL TrainingPlan type
-    const fetchedTrainingPlan: TrainingPlan = {
-        id: trainingPlan.id,
-        clientId: trainingPlan.client_id,
-        title: trainingPlan.title,
-        overview: trainingPlan.overview,
-        planJson: trainingPlan.plan_json,
-        assistantIds: trainingPlan.assistant_ids,
-        goalIds: trainingPlan.goal_ids,
-        createdAt: new Date(trainingPlan.created_at),
-        updatedAt: new Date(trainingPlan.updated_at),
-        deletedAt: trainingPlan.deleted_at ? new Date(trainingPlan.deleted_at) : null,
-        generatedBy: trainingPlan.generated_by || null,
-        sourcePrompt: trainingPlan.source_prompt || null,
-    };
-
-    console.log("Successfully fetched training plan:", fetchedTrainingPlan);
-
-    return fetchedTrainingPlan;
-};
-
-// Function to get all training plans for a specific client ID
-export const getTrainingPlansByClientId = async (clientId: string): Promise<TrainingPlan[]> => {
-    console.log("Fetching training plans for client ID:", clientId);
-
-    const { data: trainingPlans, error } = await supabaseServiceRole
-        .from('training_plans')
-        .select('*')
-        .eq('client_id', clientId);
-
-    if (error) {
-        console.error(`Error fetching training plans for client ID ${clientId}:`, error);
-        throw new Error(`Failed to fetch training plans: ${error.message}`);
-    }
-
-    // Map the database response array back to the GraphQL TrainingPlan type array
-    const fetchedTrainingPlans: TrainingPlan[] = trainingPlans.map(plan => ({
-        id: plan.id,
-        clientId: plan.client_id,
-        title: plan.title,
-        overview: plan.overview,
-        planJson: plan.plan_json,
-        assistantIds: plan.assistant_ids,
-        goalIds: plan.goal_ids,
-        createdAt: new Date(plan.created_at),
-        updatedAt: new Date(plan.updated_at),
-        deletedAt: plan.deleted_at ? new Date(plan.deleted_at) : null,
-        generatedBy: plan.generated_by || null,
-        sourcePrompt: plan.source_prompt || null,
-    }));
-
-    console.log(`Successfully fetched ${fetchedTrainingPlans.length} training plans for client ID ${clientId}.`);
-
-    return fetchedTrainingPlans;
-};
-
-// TODO: Add other repository functions (e.g., getTrainingPlan, getTrainingPlans, updateTrainingPlan, deleteTrainingPlan)
-
-// Function to get training plans for a specific user (and optionally client)
-export async function getTrainingPlans(userId: string | null, clientId: string | null): Promise<TrainingPlan[]> {
-    console.log('Fetching training plans for user:', userId, 'and client:', clientId);
-    if (!userId) {
-        console.log('getTrainingPlans: No user ID provided.');
-        return [];
-    }
-
-    let query = supabaseServiceRole
-        .from('training_plans')
-        .select('*')
-        .eq('user_id', userId); // Filter by user ID
-
-    if (clientId) {
-        query = query.eq('client_id', clientId); // Optionally filter by client ID
-    }
-
-    const { data, error } = await query;
-
-    if (error) {
-        console.error(`Error fetching training plans for user ${userId}:`, error);
-        return [];
-    }
-
-    // Map database response array to TrainingPlan type array
-    const fetchedTrainingPlans: TrainingPlan[] = data.map(plan => ({
-        ...plan,
-        createdAt: new Date(plan.created_at), // Assuming timestamp strings
-        updatedAt: new Date(plan.updated_at), // Assuming timestamp strings
-        deletedAt: plan.deleted_at ? new Date(plan.deleted_at) : null, // Handle nullable deleted_at
-    })) as TrainingPlan[]; // Type assertion
-
-    console.log(`Successfully fetched ${fetchedTrainingPlans.length} training plans for user ${userId}.`);
-    return fetchedTrainingPlans;
-}
